Change elph migrations from add to add_if_not_exists
If a project used an old version of this module it won't have the new fields. But those are needed by the built in media types. 

Just adding a new migration to your project altering the tables with the same macro as before should fix that. In case you have a new database (because you did a `mix ecto.reset` for example) this migration would then do nothing.